pyproject.toml
[project]
name = "clip-files-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.0",
    "jinja2",
    "pyyaml",
    "tiktoken",
    "pyperclip",
    "python-multipart",
    "clip-files>=0.4.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff",
    "mypy",
    "pytest>=8.3.5",
]

[tool.ruff]
line-length = 100
target-version = "py39"
exclude = ["repos", ".venv", ".git", "__pycache__"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "T20",     # flake8-print
    "ANN101",  # Missing type annotation for {name} in method
    "S101",    # Use of assert detected
    "PD901",   # df is a bad variable name. Be kinder to your future self.
    "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed in {name}
    "D402",    # First line should not be the function's signature
    "PLW0603", # Using the global statement to update `X` is discouraged
    "D401",    # First line of docstring should be in imperative mood
    "SLF001",  # Private member accessed
    "PTH",     # Use pathlib.Path
    "DTZ005",  # The use of `datetime.datetime.now()` without `tz` argument is not allowed
    "PLR0913", # Too many arguments to function call
    "S602",    # `subprocess` call with `shell=True` identified, security issue
    "S603",    # `subprocess` call: check for execution of untrusted input
    "S202",    # Uses of `tarfile.extractall()`
    "BLE001",  # Do not catch blind exception: `Exception`
    "TRY300",  # Consider moving this statement to an `else` block
    "FBT001",  # Boolean-typed positional argument in function definition
    "FBT002",  # Boolean default positional argument in function definition
    "PLR2004", # Magic value used in comparison, consider replacing `2` with a constant variable
    "FBT003",  # Boolean positional value in function call
    "E501",    # Line too long
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["SLF001"]
".github/*" = ["INP001"]

[tool.ruff.lint.mccabe]
max-complexity = 18

[tool.mypy]
python_version = "3.9"

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["repos", ".venv", ".git", "__pycache__"]
addopts = "-ra -q"
